requirements.txt
requests==2.31.0
pygame==2.5.2
geopy==2.4.1
python-dotenv==1.0.0
protobuf==4.25.1
gtfs-realtime-bindings==1.0.0
